task.py: factorize run_fuse()
There were previously some duplicated code in run_vidjil() and run_only_fuse().
This is now gathered in a unique run_fuse() function.

+++ server/web2py/applications/vidjil/models/task.py
@@ -64,7 +64,7 @@ def schedule_run(id_sequence, id_config):
     return res
 
 
-def run_vidjil(id_file, id_config, id_data, id_fuse):
+def run_vidjil(id_file, id_config, id_data, id_fuse, clean_before=False, clean_after=False):
     import time, datetime, sys, os.path
     from subprocess import Popen, PIPE, STDOUT, os
     
@@ -90,7 +90,6 @@ def run_vidjil(id_file, id_config, id_data, id_fuse):
     
     os.makedirs(out_folder)
     vidjil_log_file = open(out_folder+'/'+output_filename+'.vidjil.log', 'w')
-    fuse_log_file = open(out_folder+'/'+output_filename+'.fuse.log', 'w')
 
     ## commande complete
     cmd = DIR_VIDJIL + '/vidjil ' + ' -o  ' + out_folder + " -b " + output_filename
@@ -131,8 +130,45 @@ def run_vidjil(id_file, id_config, id_data, id_fuse):
     
     db.commit()
     
-    ## relance fuse.py 
-    output_file = out_folder+'/'+output_filename+".fused"
+    if clean_after:
+        clean_cmd = "rm -rf " + out_folder 
+        p = Popen(clean_cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
+        p.wait()
+    
+    ## l'output de Vidjil est stocké comme resultat pour l'ordonnanceur
+    ## TODO parse result success/fail
+
+    run_fuse(id_file, id_config, id_data, id_fuse, clean_before = False)
+
+    return "SUCCESS"
+
+
+
+
+
+
+
+
+def run_fuse(id_file, id_config, id_data, id_fuse, clean_before=True, clean_after=False):
+    import time, datetime, sys, os.path
+    from subprocess import Popen, PIPE, STDOUT, os
+    
+    out_folder = defs.DIR_OUT_VIDJIL_ID % id_data
+    output_filename = "%06d" % id_data
+    
+    if clean_before:
+        cmd = "rm -rf "+out_folder 
+        p = Popen(cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
+        p.wait()
+        os.makedirs(out_folder)    
+    
+    row = db(db.sequence_file.id==id_file).select()
+    id_patient = row[0].patient_id
+    
+    fuse_log_file = open(out_folder+'/'+output_filename+'.fuse.log', 'w')
+    
+    ## fuse.py 
+    output_file = out_folder+'/'+output_filename+'.fused'
     files = ""
     query = db( ( db.patient.id == db.sequence_file.patient_id )
                    & ( db.results_file.sequence_file_id == db.sequence_file.id )
@@ -141,16 +177,16 @@ def run_vidjil(id_file, id_config, id_data, id_fuse):
                    ).select( orderby=db.sequence_file.id|db.results_file.run_date, groupby=db.sequence_file.id ) 
     for row in query :
         if row.results_file.data_file is not None :
-            files += defs.DIR_RESULTS + row.results_file.data_file+" "
+            files += defs.DIR_RESULTS + row.results_file.data_file + " "
     
     cmd = "python ../fuse.py -o "+output_file+" -t 100 "+files
-    
-    
+
+
     print "=== fuse.py ==="
     print cmd
     print "==============="
     sys.stdout.flush()
-    db.commit()
+
     p = Popen(cmd, shell=True, stdin=PIPE, stdout=fuse_log_file, stderr=STDOUT, close_fds=True)
     (stdoutdata, stderrdata) = p.communicate()
     print "Output log in "+out_folder+'/'+output_filename+'.fuse.log'
@@ -164,94 +200,18 @@ def run_vidjil(id_file, id_config, id_data, id_fuse):
         raise IOError
 
     ts = time.time()
-    
-    db.fused_file[id_fuse] = dict(fuse_date = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S'),
-                                 fused_file = stream)
-    
-    db.commit()
-
-#    clean_cmd = "rm -rf " + out_folder 
-#    p = Popen(clean_cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
-#    p.wait()
-    
-    ## l'output de Vidjil est stocké comme resultat pour l'ordonnanceur
-    ## TODO parse result success/fail
-    return "SUCCESS"
-
-
-
-def run_fuse_only(id_file, id_config, id_data, id_fuse):
-    import time, datetime, sys, os.path
-    from subprocess import Popen, PIPE, STDOUT, os
-    
-    ## les chemins d'acces a vidjil / aux fichiers de sequences
-    germline_folder = DIR_VIDJIL + '/germline/'
-    upload_folder = defs.DIR_SEQUENCES
-    out_folder = defs.DIR_OUT_VIDJIL_ID % id_data
-    output_filename = "%06d" % id_data
-    
-    #clean folder
-    cmd = "rm -rf "+out_folder 
-    p = Popen(cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
-    p.wait()
-    cmd = "mkdir "+out_folder 
-    p = Popen(cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
-    p.wait()
-    
-    ## filepath du fichier input (data/clntab)
-    row = db(db.sequence_file.id==id_file).select()
-    filename = row[0].data_file
-    seq_file = upload_folder+filename
-    id_patient = row[0].patient_id
-    vidjil_germline = db.config[id_config].germline
-    
-    ## insertion dans la base de donnée
-    ts = time.time()
-    db.results_file[id_data] = dict(status = "ready",
-                                 run_date = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S')
-                                )
-    
-    ## fuse.py 
-    output_file = out_folder+'/'+output_filename+'.fused'
-    files = ""
-    query = db( ( db.patient.id == db.sequence_file.patient_id )
-                   & ( db.results_file.sequence_file_id == db.sequence_file.id )
-                   & ( db.patient.id == id_patient )
-                   & ( db.results_file.config_id == id_config )
-                   ).select( orderby=db.sequence_file.sampling_date ) 
-    for row in query :
-        if row.sequence_file.data_file is not None :
-            files += defs.DIR_RESULTS + row.sequence_file.data_file + " "
-    
-    cmd = "python ../fuse.py -o "+output_file+" -t 100 "+files
-    print cmd
-    
-    p = Popen(cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
-    p.wait()
-    print p.stdout.read()
-    
-    #store fused file
-    fuse_filepath = os.path.abspath(output_file)
-
-    try:
-        stream = open(fuse_filepath, 'rb')
-    except IOError:
-        print "!!! Fuse failed, no .fused file"
-        raise IOError
-
-    ts = time.time()
     db.fused_file[id_fuse] = dict(fuse_date = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S'),
                                  fused_file = stream)
     db.commit()
-    
-    #clean tmp folder
-    clean_cmd = "rm -rf " + out_folder 
-    p = Popen(clean_cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
-    p.wait()
+
+    if clean_after:
+        clean_cmd = "rm -rf " + out_folder 
+        p = Popen(clean_cmd, shell=True, stdin=PIPE, stdout=PIPE, stderr=STDOUT, close_fds=True)
+        p.wait()
     
     return "SUCCESS"
 
 
 from gluon.scheduler import Scheduler
 scheduler = Scheduler(db, dict(vidjil=run_vidjil,
-                               none=run_fuse_only))
+                               none=run_fuse))
